[Tests] Move more Panel test to base Properties
